Update navigation items, add pages, and separate mobile drawer from header
