+++ app/src/components/common/AuthenticatedHeader.tsx
@@ -1,20 +1,12 @@
 import * as React from 'react';
 import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
-import Divider from '@mui/material/Divider';
-import Drawer from '@mui/material/Drawer';
 import IconButton from '@mui/material/IconButton';
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemText from '@mui/material/ListItemText';
 import MenuIcon from '@mui/icons-material/Menu';
 import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
-import Button from '@mui/material/Button';
 import {Tooltip, Avatar, Menu, MenuItem, Stack} from '@mui/material';
 import logo from '../../img/favicon.png';
-import {useAuth} from '../../app/hooks/useAuth';
 import {useSignOutMutation} from '../../services/auth';
 import {useAppDispatch} from '../../app/hooks/store';
 import {setCredentials} from '../../app/authSlice';
@@ -30,57 +22,17 @@ function getInitials(name: string) {
   return splitName[0][0] + splitName[splitName.length - 1][0];
 }
 
-interface Props {
-  /**
-   * Injected by the documentation to work in an iframe.
-   * You won't need it on your project.
-   */
-  window?: () => Window;
+interface OwnProps {
+  onClick: () => void;
 }
 
-const drawerWidth = 240;
-const navItems = [
-  {title: 'Login', href: '/signin'},
-  {title: 'Sign Up', href: '/signup'},
-];
-
-export const Header = (props: Props) => {
-  const {window} = props;
-  const [mobileOpen, setMobileOpen] = React.useState(false);
-
-  const {user} = useAuth();
-
-  const handleDrawerToggle = () => {
-    setMobileOpen(prevState => !prevState);
-  };
-
-  const drawer = (
-    <Box onClick={handleDrawerToggle} sx={{textAlign: 'center'}}>
-      <Typography variant="h6" sx={{my: 2}}>
-        MUI
-      </Typography>
-      <Divider />
-      <List>
-        {navItems.map(({title}) => (
-          <ListItem key={title} disablePadding>
-            <ListItemButton sx={{textAlign: 'center'}}>
-              <ListItemText primary={title} />
-            </ListItemButton>
-          </ListItem>
-        ))}
-      </List>
-    </Box>
-  );
-
-  const container =
-    window !== undefined ? () => window().document.body : undefined;
-
+export const UserHeader = ({onClick}: OwnProps) => {
   return (
     <Box sx={{display: 'flex'}}>
       <AppBar
         position="fixed"
         sx={{
-          backgroundColor: 'white',
+          backgroundColor: '#ffffff',
           zIndex: theme => theme.zIndex.drawer + 1,
           boxShadow: 'none',
         }}
@@ -88,60 +40,23 @@ export const Header = (props: Props) => {
       >
         <Toolbar sx={{justifyContent: {xs: 'space-between'}}}>
           <IconButton
-            color="primary"
             aria-label="open drawer"
             edge="start"
-            onClick={handleDrawerToggle}
+            onClick={onClick}
             sx={{mr: 2, display: {sm: 'none'}}}
           >
             <MenuIcon />
           </IconButton>
+          <img
+            style={{width: '40px', height: '40px'}}
+            src={logo}
+            alt="Home Unite Us logo"
+          />
           <Stack direction="row" gap={1} sx={{alignItems: 'center'}}>
-            <img
-              style={{width: '40px', height: '40px'}}
-              src={logo}
-              alt="Home Unite Us logo"
-            />
-          </Stack>
-          <Stack direction="row" gap={1} sx={{alignItems: 'center'}}>
-            {!user ? (
-              <Box sx={{display: {xs: 'none', sm: 'flex'}}}>
-                {navItems.map(({title, href}) => {
-                  const variant = title === 'Sign Up' ? 'contained' : 'text';
-                  return (
-                    <Button
-                      variant={variant}
-                      href={href}
-                      color="primary"
-                      key={title}
-                    >
-                      {title}
-                    </Button>
-                  );
-                })}
-              </Box>
-            ) : null}
-            {user ? <AvatarDropdownMenu /> : null}
+            <AvatarDropdownMenu />
           </Stack>
         </Toolbar>
       </AppBar>
-      <Box component="nav">
-        <Drawer
-          container={container}
-          variant="temporary"
-          open={mobileOpen}
-          onClose={handleDrawerToggle}
-          ModalProps={{
-            keepMounted: true, // Better open performance on mobile.
-          }}
-          sx={{
-            display: {xs: 'block', sm: 'none'},
-            '& .MuiDrawer-paper': {boxSizing: 'border-box', width: drawerWidth},
-          }}
-        >
-          {drawer}
-        </Drawer>
-      </Box>
     </Box>
   );
 };
