+++ app/src/components/common/index.ts
@@ -2,3 +2,4 @@ export {Loading} from './Loading';
 export {ApplicationTrackerHeader} from './ApplicationTrackerHeader';
 export {Header} from './Header';
 export {AppLayout} from './AppLayout';
+export {UserHeader} from './AuthenticatedHeader';
