+++ app/src/views/GuestContacts.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const GuestContacts = () => {
+  return <h1>Guest Contacts</h1>;
+};
