+++ app/src/views/GuestSettings.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const GuestSettings = () => {
+  return <h1>Guest Settings</h1>;
+};
