+++ app/src/views/GuestTasks.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const GuestTasks = () => {
+  return <h1>Guest Tasks</h1>;
+};
